+++ src/main/java/com/comprehensive/eureka/recommend/service/util/UnitConverter.java
@@ -6,22 +6,22 @@ import org.springframework.stereotype.Component;
 @Component
 public class UnitConverter {
 
-    public static Integer convertToGigabytes(Integer value, String unit) {
-        return switch (unit) {
-            case "MB" -> value / 1024;
-            case "GB" -> value;
-            case "TB" -> value * 1024;
+    public static Double convertToGigabytes(Integer value, String unit) {
+        return (double) switch (unit) {
+            case "MB" -> value / 1024.0;
+            case "GB" -> (double) value;
+            case "TB" -> value * 1024.0;
             default -> throw new IllegalArgumentException("Unsupported unit: " + unit);
         };
     }
 
-    public static Integer convertToGigabytes(Integer value, String unit, DataPeriod dataPeriod) {
+    public static Double convertToGigabytes(Integer value, String unit, DataPeriod dataPeriod) {
         if (dataPeriod == DataPeriod.DAY) value *= 30;
 
-        return switch (unit) {
-            case "MB" -> value / 1024;
-            case "GB" -> value;
-            case "TB" -> value * 1024;
+        return (double) switch (unit) {
+            case "MB" -> value / 1024.0;
+            case "GB" -> (double) value;
+            case "TB" -> value * 1024.0;
             default -> throw new IllegalArgumentException("Unsupported unit: " + unit);
         };
     }
